src/requirements.txt
amqp==5.3.1
appdirs==1.4.4
asttokens==3.0.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
fonttools==4.55.8
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.0
mantis_sdk @ git+https://github.com/KellisLab/Mantis_SDK.git@01fceb6a7a1b79dfb0e12e87ba1f11d04a287bae
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
msgspec==0.19.0
nest-asyncio==1.6.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.50.0
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.1
pyppeteer==0.0.25
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
zipp==3.21.0
